+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/examples/OneMotorEncoderDrivePosition.java
@@ -16,24 +16,24 @@ public class OneMotorEncoderDrivePosition extends LinearOpMode {
 
     // Data for GoBilda 312 RPM motor
     private final int maxMotorRPM = 6000;
-    private final float gearRatio = 19.2f;
+    private final double gearRatio = 19.2f;
     private final int ticksPerMotorRev = 28;
-    private final float ticksPerWheelRev = ticksPerMotorRev * gearRatio;
+    private final double ticksPerWheelRev = ticksPerMotorRev * gearRatio;
 
     // GoBilda mecanum wheels - grippy rollers - 104mm diameter
     // GoBilda mecanum wheels - smooth rollers - 96 mm diameter
-    private final float wheelDiameterInches = 104.0f / 25.4f;
-    private final float wheelCircumferenceInches = wheelDiameterInches * 3.1415f;
+    private final double wheelDiameterInches = 104.0f / 25.4f;
+    private final double wheelCircumferenceInches = wheelDiameterInches * 3.1415f;
 
     // Determine how many ticks required to move forward one inch.
-    private final float ticksPerInch = ticksPerWheelRev / wheelCircumferenceInches;
+    private final double ticksPerInch = ticksPerWheelRev / wheelCircumferenceInches;
 
     // Target position relative to start
-    private float targetInches = 0.0f;
+    private double targetInches = 0.0f;
 
-    private final float targetInchesIncrement = 12.0f;
+    private final double targetInchesIncrement = 12.0f;
     private int targetTicks = 0;
-    private float motorPower = 0.5f;
+    private double motorPower = 0.5f;
 
     @Override
     public void runOpMode() {
@@ -54,7 +54,7 @@ public class OneMotorEncoderDrivePosition extends LinearOpMode {
             } else if (gamepad1.a) {
                 targetInches -= targetInchesIncrement;
             }
-            targetTicks = Math.round(targetInches * ticksPerInch);
+            targetTicks = Math.toIntExact(Math.round(targetInches * ticksPerInch));
             motor.setTargetPosition(targetTicks);
 
             telemetry.addData("targetInches: ", targetInches);
